% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_slides.R
\name{browse_slides}
\alias{browse_slides}
\title{Open rendered, hostes slides}
\usage{
browse_slides(.module)
}
\arguments{
\item{.module}{the name of the module}
}
\description{
Open rendered, hostes slides
}
