% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_solutions.R
\name{check_solutions}
\alias{check_solutions}
\alias{check_shiny_solutions}
\title{Check solutions}
\usage{
check_solutions(.file = NULL)

check_shiny_solutions(.file = NULL)
}
\arguments{
\item{.file}{The path of the files to check. If \code{NULL}, searches directory
for relevant files.}
}
\description{
Check solutions hosted in \code{emptyfield-ds/solutions}
}
