% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module.R
\name{create_module}
\alias{create_module}
\title{Create a new teaching module}
\usage{
create_module(
  module_dir,
  module_title,
  module_subtitle = "",
  include_exercises = TRUE,
  include_solutions = TRUE,
  include_img = TRUE
)
}
\arguments{
\item{module_dir}{The directory name. Also the file name of the slides file.}

\item{module_title}{The title of the module}

\item{module_subtitle}{The subtitle of the module}

\item{include_exercises}{Include \code{exercises.Rmd}?}

\item{include_solutions}{Include \code{solutions.Rmd}?}

\item{include_img}{Include an \verb{img/} folder?}
}
\value{
\code{module_dir}, invisibly.
}
\description{
Create a new directory with R Markdown files for slides, exercises, and solutions.
}
