% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_pkg.R
\name{dehydrate_rproj}
\alias{dehydrate_rproj}
\alias{rehydrate_rproj}
\title{Dehydrate and rehydrate rproj files for safe project hygiene}
\usage{
dehydrate_rproj(dir)

rehydrate_rproj(dir)
}
\arguments{
\item{x}{the Rproj file}
}
\value{
a filename of either \code{.Rproj} or \code{.dehydrated_Rproj}
}
\description{
Dehydrate and rehydrate rproj files for safe project hygiene
}
