% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_r_pkg_module.R
\name{open_r_pkg_module}
\alias{open_r_pkg_module}
\title{Copy a teaching package to a temporary directory and open it in RStudio}
\usage{
open_r_pkg_module(
  path = path_warehouse(),
  module = c("setup", "write_code", "document", "test", "teach", "add_files"),
  open = interactive()
)
}
\arguments{
\item{path}{The root path where the teaching material is}

\item{module}{The R Package teaching module to open}

\item{open}{Open in RStudio?}
}
\value{
Invisibly, the temporary directory where the module package has been copied
}
\description{
Copy a teaching package to a temporary directory and open it in RStudio
}
