% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_temp_module.R
\name{open_temp_module}
\alias{open_temp_module}
\title{Copy a teaching package to a temporary directory and open it in RStudio}
\usage{
open_temp_module(module, github = TRUE, open = interactive())
}
\arguments{
\item{module}{The path to a local module or the name of a repo on
\code{emptyfile-ds}}

\item{github}{Use the copy of the module on GitHub?}

\item{open}{Open in RStudio?}
}
\value{
Invisibly, the temporary directory where the module package has been
copied
}
\description{
Copy a teaching package to a temporary directory and open it in RStudio
}
