% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_module.R
\name{path_warehouse}
\alias{path_warehouse}
\title{Write a path to the teaching warehouse}
\usage{
path_warehouse(...)
}
\arguments{
\item{...}{Additional path details passed to \code{\link[fs:path]{fs::path()}}.}
}
\value{
A path
}
\description{
\code{path_warehouse()} uses the \code{shunyata.teaching_warehouse} option to assemble
an path to a given module.
}
\examples{
path_warehouse("purrr_basics")
}
