% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_cloud.R
\name{rsc_archive_course}
\alias{rsc_archive_course}
\title{Archive an RStudio Cloud Course}
\usage{
rsc_archive_course(
  space_id,
  rsc_user_id = getOption("shunyata.rscloud_user_id")
)
}
\arguments{
\item{space_id}{The RStudio Cloud space ID}

\item{rsc_user_id}{The user ID of the course lead}
}
\value{
Invisibly, the course ID
}
\description{
\code{rsc_archive_course()} removes members except for \code{rsc_user_id}, moving space
member projects to their own workspace.
}
