% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module.R
\name{sanitize_module}
\alias{sanitize_module}
\title{Sanitize module}
\usage{
sanitize_module(
  module,
  new_path = NULL,
  rm_files = c("theme.css", "kakashi.css", "solutions.Rmd"),
  rm_dir = c("img", "libs"),
  overwrite = FALSE
)
}
\arguments{
\item{module}{The name of the module}

\item{new_path}{Where to copy the sanitized directory}

\item{rm_files}{Which files to remove}

\item{rm_dir}{Which directories to remove}

\item{overwrite}{Overwrite the sanitized module if it already exists?}
}
\value{
invisibly, the path of the new module
}
\description{
Remove files and directories that aren't needed for the student copy
}
