% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shunyata-package.R
\docType{package}
\name{shunyata-package}
\alias{shunyata}
\alias{shunyata-package}
\title{shunyata: Deploy Empty Field Teaching Material}
\description{
This package helps set up new teaching material and deploy courses based on the Empty Field teaching workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emptyfield-ds/shunyata}
  \item Report bugs at \url{https://github.com/emptyfield-ds/shunyata/issues}
}

}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})

}
\keyword{internal}
