% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_module.R
\name{sync_module}
\alias{sync_module}
\title{Sync a module to GitHub}
\usage{
sync_module(.dir)
}
\arguments{
\item{.dir}{The path to the module}
}
\description{
\code{sync_module()} takes a raw teaching module, renders slides, sanitizes, and
syncs with the existing GitHub repo for the module. Use with
\code{\link[=path_warehouse]{path_warehouse()}}.
}
