% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_pkg.R
\name{temp_pkg}
\alias{temp_pkg}
\title{Copy a teaching package to a temporary directory and open it in RStudio}
\usage{
temp_pkg(
  path = ".",
  module = c("setup", "write_code", "document", "test", "teach", "add_files"),
  open = interactive()
)
}
\arguments{
\item{path}{The root path where the teaching material is}

\item{module}{The R Package teaching module to open}

\item{open}{Open in RStudio?}
}
\value{
Invisibly, the temporary directory where the module package has been copied
}
\description{
Copy a teaching package to a temporary directory and open it in RStudio
}
