% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_course_template.R
\name{use_course_template}
\alias{use_course_template}
\title{Use course template}
\usage{
use_course_template(
  template,
  save_as = template,
  data = list(),
  ignore = FALSE,
  open = FALSE
)
}
\arguments{
\item{template}{Path to template file relative to \verb{templates/} directory
within \code{package}; see details.}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{data}{A list of data passed to the template.}

\item{ignore}{Should the newly created file be added to \code{.Rbuildignore}?}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
A wrapper around \link[usethis:use_template]{usethis::use_template} to use templates in this package.
}
