% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license.R
\name{use_edu_cca_license}
\alias{use_edu_cca_license}
\title{Add CC-by-A License to Repo \code{README}}
\usage{
use_edu_cca_license(proj = usethis::proj_get())
}
\arguments{
\item{proj}{The directory of the project, by default the usethis project.}
}
\value{
Invisibly, the path to the \code{README}
}
\description{
\code{use_edu_cca_license()} adds a Creative Commons Attribution 4.0 International
License to the \code{README} of an educational repository.
}
