% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ef_github.R
\name{use_ef_github}
\alias{use_ef_github}
\title{Create a Bare-Bones Repo on Empty Field DS}
\usage{
use_ef_github(repo_name, private = FALSE)
}
\arguments{
\item{repo_name}{The name of the repo to be created}

\item{private}{Logical. Should the repo be private?}
}
\value{
Opens the URL of the repo
}
\description{
\code{use_ef_github()} creates a bare-bones repo intended for use with
\code{\link[=sync_module]{sync_module()}}
}
