% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_deps.R
\name{write_deps}
\alias{write_deps}
\alias{detect_deps}
\alias{read_deps}
\title{Manage module dependencies}
\usage{
write_deps(deps, path = ".")

detect_deps(path)

read_deps(modules)
}
\arguments{
\item{deps}{A vector of dependencies}

\item{path}{A path to the module}

\item{modules}{A vector of modules}
}
\value{
A vector of module dependencies
}
\description{
\code{read_deps()} pulls the package dependencies in \code{.deps} for \code{modules} from GitHub
}
