% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_cloud.R
\name{zip_modules}
\alias{zip_modules}
\title{Zip modules to upload to RStudio Cloud}
\usage{
zip_modules()
}
\value{
invisibly, a vector of directories
}
\description{
\code{zip_modules()} zips modules in a directory for upload to RStudio Cloud, e.g.
the \code{mastering_r_for_epi} repository.
}
