% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_temp_module.R
\name{open_temp_module}
\alias{open_temp_module}
\title{Copy a teaching package to a temporary directory and open it in RStudio}
\usage{
open_temp_module(module, github = TRUE, open = interactive())
}
\arguments{
\item{module}{The R Package teaching module to open}

\item{open}{Open in RStudio?}

\item{path}{The root path where the teaching material is}
}
\value{
Invisibly, the temporary directory where the module package has been copied
}
\description{
Copy a teaching package to a temporary directory and open it in RStudio
}
