% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_slides.R
\name{render_slides}
\alias{render_slides}
\title{Render Module Slides}
\usage{
render_slides(dir, qmd_name = NULL, render = TRUE)
}
\arguments{
\item{dir}{The directory where the module is located}

\item{qmd_name}{The name of Quarto file. By default, assumes that it is the same
as \code{dir} with a \code{.qmd} extension.}

\item{render}{Logical. Render the Quarto file?}
}
\value{
Invisibly, the path to the rendered PDF
}
\description{
Render Module Slides
}
