% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_module.R
\name{sanitize_module}
\alias{sanitize_module}
\title{Sanitize module}
\usage{
sanitize_module(
  module,
  new_path = NULL,
  rm_files = c("theme.css", "kakashi.css", "solutions.Rmd"),
  rm_dir = c("img", "libs"),
  overwrite = FALSE,
  build_slides = TRUE,
  rmd_name = NULL,
  ...
)
}
\arguments{
\item{module}{The name of the module}

\item{new_path}{Where to copy the sanitized directory}

\item{rm_files}{Which files to remove}

\item{rm_dir}{Which directories to remove}

\item{overwrite}{Overwrite the sanitized module if it already exists?}

\item{build_slides}{Logical. Render the slides?}

\item{rmd_name}{The name of Rmd file. By default, assumes that it is the same
as \code{dir} with a \code{.Rmd} extension.}
}
\value{
invisibly, the path of the new module
}
\description{
Remove files and directories that aren't needed for the student copy
}
